requirements.txt
Django==4.2.1
psycopg2-binary==2.9.6
django-extensions==3.2.1
ipython==8.14.0
aiogram==2.25.1
python-dotenv==1.0.0
asyncpg==0.27.0
SQLAlchemy==1.3.24

